feat: rewrite modules/cjmToEsm
